move git source-related code to pl.nort.config.source.git package
