+++ src/main/java/pl/nort/config/source/git/GitConfigurationSourceException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package pl.nort.config.source;
+package pl.nort.config.source.git;
 
 public class GitConfigurationSourceException extends RuntimeException {
