+++ src/test/java/pl/nort/config/source/git/GitConfigurationSourceIntegrationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package pl.nort.config.source;
+package pl.nort.config.source.git;
 
 import static org.assertj.core.api.Assertions.assertThat;
